+++ src/main/java/kr/codesqaud/cafe/app/user/controller/UserController.java
@@ -4,10 +4,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import kr.codesqaud.cafe.app.user.controller.dto.UserLoginRequest;
+import kr.codesqaud.cafe.app.user.controller.dto.UserModifiedResponse;
 import kr.codesqaud.cafe.app.user.controller.dto.UserResponse;
 import kr.codesqaud.cafe.app.user.controller.dto.UserSavedRequest;
-import kr.codesqaud.cafe.app.user.controller.dto.UserModifiedResponse;
 import kr.codesqaud.cafe.app.user.entity.User;
 import kr.codesqaud.cafe.app.user.service.UserService;
 import org.slf4j.Logger;
@@ -19,7 +18,6 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.view.RedirectView;
 
 @RestController
 public class UserController {
@@ -71,31 +69,6 @@ public class UserController {
         return new UserResponse(user);
     }
 
-    // 로그인
-    @PostMapping("/users/login")
-    public ModelAndView login(@Valid @RequestBody UserLoginRequest requestDto,
-        HttpSession session) {
-        logger.info("login" + requestDto.toString());
-        User user = userService.login(requestDto);
-        session.setAttribute("user", new UserResponse(user));
-        return new ModelAndView(new RedirectView("/"));
-    }
-
-    // 로그아웃
-    @PostMapping("/users/logout")
-    public ModelAndView logout(HttpSession session) {
-        ModelAndView mav = new ModelAndView("user/login");
-        session.removeAttribute("user");
-        return mav;
-    }
-
-    // 비밀번호 확인
-    @PostMapping("/users/password/{id}")
-    public ModelAndView passwordConfirm(@PathVariable(value = "id") Long id,
-        @Valid @RequestBody UserSavedRequest requestDto) {
-        userService.confirmPassword(id, requestDto);
-        return new ModelAndView(new RedirectView("/user/form/" + id));
-    }
 
     // 회원가입 페이지
     @GetMapping("/user/form")
@@ -113,17 +86,5 @@ public class UserController {
         return mav;
     }
 
-    // 로그인 페이지
-    @GetMapping("/user/login")
-    public ModelAndView loginForm() {
-        return new ModelAndView("user/login");
-    }
 
-    // 비밀번호 확인 페이지
-    @GetMapping("/user/password/{id}")
-    public ModelAndView passwordConfirmForm(@PathVariable(value = "id") Long id) {
-        ModelAndView mav = new ModelAndView("user/passwordForm");
-        mav.addObject("user", userService.findUser(id));
-        return mav;
-    }
 }
